Refactor generateArray tool and remove unused createArtistSegments tool
