+++ lib/tools/getMcpTools.ts
@@ -36,7 +36,6 @@ import getScheduledActions from "./scheduled_actions/getScheduledActions";
 import updateScheduledAction from "./scheduled_actions/updateScheduledActions";
 import deleteScheduledAction from "./scheduled_actions/deleteScheduledActions";
 import getSocialFans from "./getSocialFans";
-import createArtistSegments from "./createArtistSegments";
 
 export async function getMcpTools() {
   const tools = {
@@ -78,7 +77,6 @@ export async function getMcpTools() {
     update_scheduled_action: updateScheduledAction,
     delete_scheduled_actions: deleteScheduledAction,
     get_social_fans: getSocialFans,
-    create_artist_segments: createArtistSegments,
   };
 
   return tools;
